RF | Combining FooterTest classes US 13.01 (#2199)
* RF | Combining FooterTest classes US 13.01

* RF | Combining FooterTest classes US 13.01

* RF | Combining FooterTest classes US 13.01
