+++ src/test/java/school/redrover/FooterJenkinsVersionTest.java
@@ -2,17 +2,16 @@ package school.redrover;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.model.JenkinsVersionPage;
 import school.redrover.model.MainPage;
+import school.redrover.model.ManageJenkinsPage;
 import school.redrover.runner.BaseTest;
 
 
 public class FooterJenkinsVersionTest extends BaseTest {
-    private final String expectedJenkinsVersion = "Jenkins 2.387.2";
-    private final String expectedSiteTitle = "Jenkins";
 
     @Test
     public void testFooterJenkinsVersion() {
@@ -33,21 +32,53 @@ public class FooterJenkinsVersionTest extends BaseTest {
         WebElement jenkinsVersion = getDriver().findElement(By.xpath("//a[@target='_blank']"));
         String actualJenkinsVersion = jenkinsVersion.getText();
 
+        String expectedJenkinsVersion = "Jenkins 2.387.2";
         Assert.assertEquals(actualJenkinsVersion, expectedJenkinsVersion, "Jenkins version does not match");
     }
 
-    @Ignore
     @Test
-    public void testClickOnJenkinsVersionOpensSiteOnNodesPage(){
+    public void testCheckJenkinsVersionInPeoplePage(){
+        WebElement peoplePageButton = getDriver().findElement(By.xpath("//a[@href='/asynchPeople/']"));
+        peoplePageButton.click();
 
-        getDriver().findElement(By.xpath("//span[@class='pane-header-title']/a")).click();
-        getDriver().findElement(By.xpath("//a[@target='_blank']")).click();
+        getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//h1[contains(text(),'People')]")));
 
-        for(String winHandle : getDriver().getWindowHandles()) {
-            getDriver().switchTo().window(winHandle);
-        }
-
-        String actualSiteTitle = getDriver().findElement(By.xpath("//h1[@class='page-title']/span")).getText();
-        Assert.assertEquals(actualSiteTitle, expectedSiteTitle);
+        Assert.assertEquals(getDriver().findElement(By.xpath("//a[@rel='noopener noreferrer']")).getText(),"Jenkins 2.387.2");
     }
+
+    @Test
+    public void testVerifyJenkinsVersionOnManageJenkinsPage() {
+        ManageJenkinsPage manageJenkinsPage = new ManageJenkinsPage(getDriver())
+                .navigateToManageJenkinsPage()
+                .scrollToFooterPageByJenkinsVersionBTN();
+
+        Assert.assertTrue(manageJenkinsPage.getVersionJenkinsFromFooter(), "Wrong version Jenkins");
+    }
+
+
+    private static final String version = "Jenkins 2.387.2";
+    private static final By VERSION_NUMBER = By.xpath("//a[@href = 'https://www.jenkins.io/']");
+
+    public void assertVersion () {
+        Assert.assertEquals(getDriver().findElement(VERSION_NUMBER).getText(), version);
+    }
+
+    @Test
+    public void testVersionOnBuildHistoryPage () {
+        getDriver().findElement(By.xpath("//*[@id = 'tasks']/div[3]//a")).click();
+
+        assertVersion();
+    }
+
+    @Test
+    public void testVersionOnMyViewPage () {
+        getDriver().findElement(By.xpath("//*[@id = 'tasks']/div[5]//a")).click();
+
+        assertVersion();
+    }
+
+
+
+
+
 }
