+++ src/test/java/school/redrover/FreestyleProject4Test.java
@@ -5,6 +5,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
+import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.runner.BaseTest;
 import school.redrover.runner.TestUtils;
@@ -63,6 +64,7 @@ public class FreestyleProject4Test extends BaseTest {
         Assert.assertEquals(actualBuildStatus, expectedBuildStatus);
     }
 
+    @Ignore
     @Test(dependsOnMethods = "testBuildNowProject")
     public void testBuildNowProjectWithBooleanParameter() {
         new Actions(getDriver())
